// src/app/domain/models/doctors.model.ts
export interface DoctorsModel {
    id?: number
    cmp?: number
    full_name_doctor?: string
    code_sap?: string
    category?: string
    local_address?: string
    local_district?: string
    code_closeup?: string
    specialty?: string
    previous_contract_end_date?: string
    renovation?: number
}